Collapse verify-corpus into one file and add debug entrypoint
Merges the generator body and descriptor into a single
packages/spec/workflows/verify-corpus.ts, relying on the compiler's
call-graph reachability to keep descriptor-side @tisyn/config imports
unreachable from verifyCorpus. Adds a named `debug` entrypoint that
swaps the default memory journal for a file-backed NDJSON journal at
./.debug/verify-corpus.ndjson (env-overridable via
TISYN_VERIFY_CORPUS_JOURNAL), ignored via .gitignore.

The e2e test gains a debug-path case asserting the journal file
exists, parses as NDJSON, and terminates with a close event. Both
cases remove the journal file before spawning so ordering is
immaterial and the default-path "no file written" assertion is
hermetic.
